requirements.txt
streamlit
pandas
python-docx
pdfplumber
sentence-transformers
scikit-learn
matplotlib